requirements.txt
asgiref==3.5.2
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
distlib==0.3.2
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-rest-knox==4.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.13.1
filelock==3.0.12
platformdirs==2.0.2
pycparser==2.21
pytz==2022.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.6.0
